Gate the follower DLL on matching and fix the delay and timer edges.
The follower servo only runs (and relocks on engagement) when actually
rate-matching a foreign clock; published delay scales device frames but
not the graph-side resampler queue; the idle tick advances next_time;
invokes report -EIO when the closure didn't run and clear() aborts
rather than freeing state under a live source; the timer rearms only on
a real io or role transition (resetting its phase on every re-point
bunched cycles during renegotiation storms).

+++ src/sound.rs
@@ -207,7 +207,7 @@ fn set_value(fd: c_int, req: c_ulong, value: u32, tolerance: u32) -> Result<(), 
   if unsafe { libc::ioctl(fd, req, &mut v) } == -1 {
     return Err(Errno::last());
   }
-  if (v - value as c_int).unsigned_abs() > tolerance {
+  if (v as i64 - value as i64).unsigned_abs() > tolerance as u64 {
     return Err(Errno::EINVAL);
   }
   Ok(())
